+++ src/1_Page/Profile/index.js
@@ -1,8 +1,7 @@
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import STYLE from "./style";
 
 import Header from "./ui/Header";
-import TrackingImageTab from "./ui/TrackingImageTab";
 
 import useTabs from "./model/useTabs";
 import useSettingMode from "./model/useSettingMode";
@@ -11,13 +10,15 @@ import useManageTrackData from "./model/useManageTrackData.js";
 
 import useGetProfileTrackingImageList from "../../3_Entity/Tracking/useGetProfileTrackingImageList.js";
 import { useState } from "react";
+import ModifyModeHeader from "./ui/ModifyModeHeader/index.js";
+import TrackingImageTabContainer from "./ui/TrackingImageTabContainer/index.js";
 
 let profile = 0;
 const Profile = () => {
   const { userIdx } = useParams();
 
   const [tabState, handleTabClick] = useTabs(); // 탭 관리 훅
-  const [modifyMode, memoizedSetMode] = useSettingMode(); // 수정 , 삭제 상태 관리
+  const [modifyMode, handleSetMode, handleCloseMode] = useSettingMode(); // 수정 , 삭제 상태 관리
   const [shareTabPage, shareObserveRef] = useInfinityScroll();
   const [saveTabPage, saveObserveRef] = useInfinityScroll();
 
@@ -39,38 +40,40 @@ const Profile = () => {
 
   profile += 1;
   console.log("프로필", profile);
+  console.log(displayTrackingImage);
+
   return (
     <>
       <STYLE.Main>
-        <Header
-          setMode={memoizedSetMode}
-          activeTabStr={tabState?.activeTabStr}
-          handleTabClick={handleTabClick}
-          modifyIdxList={modifyIdxList}
-          displayTrackingImage={displayTrackingImage}
-          setDisplayTrackingImage={setDisplayTrackingImage}
-          setModifyIdxList={setModifyIdxList}
-          backupTrackingImageData={backupTrackingImageData}
-        />
-        <STYLE.SliderWrapper>
-          <STYLE.Slider $tabIndex={tabState?.tabIndex}>
-            <TrackingImageTab
-              displayTrackingImage={displayTrackingImage.share}
-              modifyMode={modifyMode}
-              obServeRef={hasMoreContent.share ? shareObserveRef : null}
-              setDisplayTrackingImage={setDisplayTrackingImage}
-              setModifyIdxList={setModifyIdxList}
-            />
-            <TrackingImageTab
-              displayTrackingImage={displayTrackingImage.save}
-              modifyMode={modifyMode}
-              obServeRef={hasMoreContent.save ? saveObserveRef : null}
-              setDisplayTrackingImage={setDisplayTrackingImage}
-              setModifyIdxList={setModifyIdxList}
-            />
-          </STYLE.Slider>
-        </STYLE.SliderWrapper>
+        {modifyMode ? (
+          <ModifyModeHeader
+            modifyMode={modifyMode}
+            handleCloseMode={handleCloseMode}
+            modifyIdxList={modifyIdxList}
+            setDisplayTrackingImage={setDisplayTrackingImage}
+            setModifyIdxList={setModifyIdxList}
+            backupTrackingImageData={backupTrackingImageData}
+          />
+        ) : (
+          <Header
+            handleSetMode={handleSetMode}
+            handleTabClick={handleTabClick}
+            activeTabStr={tabState?.activeTabStr}
+          />
+        )}
       </STYLE.Main>
+
+      <TrackingImageTabContainer
+        modifyMode={modifyMode}
+        hasMoreContent={hasMoreContent}
+        tabIndex={tabState.tabIndex}
+        handleTabClick={handleTabClick}
+        displayTrackingImage={displayTrackingImage}
+        setDisplayTrackingImage={setDisplayTrackingImage}
+        shareObserveRef={shareObserveRef}
+        setModifyIdxList={setModifyIdxList}
+        saveObserveRef={saveObserveRef}
+      />
 
       {loading && (
         <STYLE.LoadingContainer>
