+++ src/1_Page/Profile/lib/profileUtil.js
@@ -42,13 +42,3 @@ export const removeDuplicateData = (data) => {
     (item, index, self) => index === self.findIndex((t) => t.idx === item.idx)
   );
 };
-
-// idx를 기준으로 오름차순 정렬
-export const sortTrackData = (data) => {
-  return data.sort((a, b) => b.idx - a.idx);
-};
-
-export const moveTrack = (track, fromList, toList, sharingStatus) => ({
-  updatedFrom: fromList.filter((item) => item.idx !== track.idx),
-  updatedTo: [...toList, { ...track, sharing: sharingStatus }],
-});
