+++ src/1_Page/Profile/ui/TrackingImageTabContainer/ui/TrackingImageBox/model/useSelectTrack.js
@@ -1,4 +1,4 @@
-import { moveTrack, sortTrackData } from "../../../../../lib/profileUtil";
+import { moveTrack, sortTrackData } from "../lib/profileUtil";
 
 const useSelectTrack = (setDisplayTrackingImage, setModifyIdxList) => {
   // 공유 상태 이벤트일 경우 트랙 위치 토글
