+++ src/1_Page/Profile/ui/TrackingImageTabContainer/ui/TrackingImageBox/style.js
